requirements.txt
azure-mgmt-storage
azure-mgmt-resource

